requirements.txt
altair==5.1.2
annotated-types==0.6.0
app-store-reviews-reader==1.2
atlassian-python-api==3.41.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
blinker==1.6.3
blis==0.7.11
bs4==0.0.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.5.14
dateparser==1.1.8
Deprecated==1.2.14
dnspython==1.16.0
elastic-transport==8.10.0
elasticsearch==8.10.1
feedparser==6.0.10
filelock==3.13.0
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gnews==0.3.5
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-play-scraper==1.2.4
googleapis-common-protos==1.61.0
httplib2==0.22.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
mdurl==0.1.2
missingno==0.5.2
mmh3==4.0.1
mpmath==1.3.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
obsei==0.0.14
packaging==23.2
pandas==1.5.3
phonenumbers==8.13.23
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
praw==7.7.1
prawcore==2.4.0
preshed==3.0.9
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.19.0
pydantic==1.10.13
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pymongo==3.12.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-facebook-api==0.15.0
pytz==2023.3.post1
PyYAML==6.0.1
reddit-rss-reader==1.3.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
searchtweets-v2==1.1.1
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
sklearn-pandas==2.2.0
slack-sdk==3.23.0
smart-open==6.4.0
smmap==5.0.1
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
srsly==2.4.8
streamlit==1.28.0
sympy==1.12
tenacity==8.2.3
textblob==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tldextract==5.0.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.18
vaderSentiment==3.3.2
validators==0.22.0
wasabi==1.1.2
weasel==0.3.3
websocket-client==1.6.4
wordcloud==1.9.2
wrapt==1.15.0
zipp==3.17.0
